+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### `testlib`
 
+* Add `fullAnnBinarySpec` and `fullNoAnnBinarySpec` test harnesses to `Test.Cardano.Ledger.Core.Binary`
 * Add `TestBlockHeader` and `mkTestBlockHeaderNoNonce` as a replacement to deprecated `BHeaderView` and `makeHeaderView`.
 * Modify `ToExpr` instance for `Mismatch` to display type-level `r` parameter using `Typeable`